+++ src/main/java/com/example/ArtAuction_24/security/CustomOAuth2User.java
@@ -1,0 +1,25 @@
+package com.example.ArtAuction_24.security;
+
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.oauth2.core.user.OAuth2User;
+
+import java.util.Collection;
+import java.util.Map;
+
+class CustomOAuth2User extends User implements OAuth2User {
+
+    public CustomOAuth2User(String username, String password, Collection<? extends GrantedAuthority> authorities) {
+        super(username, password, authorities);
+    }
+
+    @Override
+    public Map<String, Object> getAttributes() {
+        return null;
+    }
+
+    @Override
+    public String getName() {
+        return getUsername();
+    }
+}
