+++ src/main/java/com/example/ArtAuction_24/security/CustomOAuth2UserService.java
@@ -48,19 +48,3 @@ public class CustomOAuth2UserService extends DefaultOAuth2UserService {
     }
 }
 
-class CustomOAuth2User extends User implements OAuth2User {
-
-    public CustomOAuth2User(String username, String password, Collection<? extends GrantedAuthority> authorities) {
-        super(username, password, authorities);
-    }
-
-    @Override
-    public Map<String, Object> getAttributes() {
-        return null;
-    }
-
-    @Override
-    public String getName() {
-        return getUsername();
-    }
-}
